Conflicts:
	modules/cms-core/pom.xml
	modules/cms-core/src/main/java/com/enonic/cms/core/content/ContentStorer.java
	modules/cms-core/src/main/java/com/enonic/cms/core/content/category/UnitEntity.java
	modules/cms-core/src/main/java/com/enonic/cms/core/portal/image/ImageServiceImpl.java
	modules/cms-core/src/main/java/com/enonic/vertical/adminweb/ContentTypeHandlerServlet.java
	modules/cms-core/src/main/java/com/enonic/vertical/adminweb/UserStoreHandlerServlet.java
	modules/cms-core/src/main/java/com/enonic/vertical/engine/handlers/ContentHandler.java
	modules/cms-core/src/main/resources/META-INF/spring/core.xml
	modules/cms-core/src/test/java/com/enonic/cms/core/content/contenttype/ContentTypeImportConfigParserTest.java
	modules/cms-core/src/test/java/com/enonic/cms/core/plugin/config/PluginConfigHelperTest.java
	modules/cms-core/src/test/java/com/enonic/cms/core/security/userstore/GroupMembershipDiffResolverTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/client/AbstractInternalClientImpl_ImportContentTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/client/InternalClientImpl_AssignContentTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/client/InternalClientImpl_SnapshotContentTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/client/InternalClientImpl_UpdateContentTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/client/InternalClientImpl_getBinaryTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/client/InternalClientImpl_getContentTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/content/ContentAccessResolverTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/content/ContentServiceImplTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/content/ContentServiceImpl_accessTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/content/ContentServiceImpl_assignContentTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/content/ContentServiceImpl_contentConfigTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/content/ContentServiceImpl_copyContentTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/content/ContentServiceImpl_createContentTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/content/ContentServiceImpl_queryContentTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/content/ContentServiceImpl_relatedcontentTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/content/ContentServiceImpl_snapshotContentTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/content/ContentServiceImpl_storeRelatedContentTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/content/ContentServiceImpl_updateContentStatusTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/content/ContentServiceImpl_updateContentTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/content/ContentServiceImpl_withFileHandlerTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/content/ContentServiceImpl_withImageHandlerTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/datasources/DataSourceServiceImpl_getContentByCategory_queryTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/datasources/DataSourceServiceImpl_getContentByQueryTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/datasources/DataSourceServiceImpl_getContentByQuery_relatedContentTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/datasources/DataSourceServiceImpl_getRelatedContentTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/datasources/DatasourceServiceImpl_getContentTest.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/plugin/MockPluginManager.java
	modules/cms-itest/src/test/java/com/enonic/cms/itest/structure/menuitem/MenuItemServiceImplTest.java
	pom.xml

+++ modules/cms-core/src/main/java/com/enonic/cms/core/structure/menuitem/MenuItemComparatorByHierarchy.java
@@ -43,7 +43,7 @@ public class MenuItemComparatorByHierarchy
         }
 
         // NOTE: This has to be done to prevent trouble with equals when menuItems have same order but different keys
-        return ( a.getKey() < b.getKey() ? -1 : ( a.getKey() == b.getKey() ? 0 : 1 ) );
+        return ( a.getKey().toInt() < b.getKey().toInt() ? -1 : ( a.getKey().toInt() == b.getKey().toInt() ? 0 : 1 ) );
     }
