+++ modules/cms-core/src/main/java/com/enonic/vertical/engine/handlers/SectionHandler.java
@@ -4,29 +4,41 @@
  */
 package com.enonic.vertical.engine.handlers;
 
-import com.enonic.cms.core.SiteKey;
-import com.enonic.cms.core.security.user.User;
-import com.enonic.cms.core.structure.menuitem.MenuItemEntity;
-import com.enonic.cms.core.structure.menuitem.MenuItemKey;
-import com.enonic.cms.framework.util.TIntArrayList;
-import com.enonic.cms.framework.xml.XMLDocument;
-import com.enonic.cms.framework.xml.XMLDocumentFactory;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Timestamp;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import org.w3c.dom.Document;
+import org.w3c.dom.Element;
+
 import com.enonic.esl.sql.model.Column;
 import com.enonic.esl.xml.XMLTool;
-import com.enonic.vertical.engine.*;
+import com.enonic.vertical.engine.SectionCriteria;
+import com.enonic.vertical.engine.VerticalCreateException;
+import com.enonic.vertical.engine.VerticalEngineLogger;
+import com.enonic.vertical.engine.VerticalRemoveException;
+import com.enonic.vertical.engine.VerticalSecurityException;
+import com.enonic.vertical.engine.XDG;
 import com.enonic.vertical.engine.dbmodel.ContentMinimalView;
 import com.enonic.vertical.engine.dbmodel.ContentView;
 import com.enonic.vertical.engine.dbmodel.SectionView;
 import com.enonic.vertical.engine.processors.AttributeElementProcessor;
 import com.enonic.vertical.engine.processors.ElementProcessor;
-import org.w3c.dom.Document;
-import org.w3c.dom.Element;
 
-import java.sql.*;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import com.enonic.cms.framework.util.TIntArrayList;
+import com.enonic.cms.framework.xml.XMLDocument;
+import com.enonic.cms.framework.xml.XMLDocumentFactory;
+
+import com.enonic.cms.core.SiteKey;
+import com.enonic.cms.core.security.user.User;
+import com.enonic.cms.core.structure.menuitem.MenuItemEntity;
+import com.enonic.cms.core.structure.menuitem.MenuItemKey;
 
 public class SectionHandler
     extends BaseHandler
@@ -286,9 +298,9 @@ public class SectionHandler
         if ( entity != null && entity.isSection() )
         {
             Element elem = doc.getDocumentElement();
-            elem.setAttribute( "key", String.valueOf( entity.getKey() ) );
+            elem.setAttribute( "key", String.valueOf( entity.getKey().toInt() ) );
             elem.setAttribute( "menukey", String.valueOf( entity.getSite().getKey() ) );
-            elem.setAttribute( "menuitemkey", String.valueOf( entity.getKey() ) );
+            elem.setAttribute( "menuitemkey", String.valueOf( entity.getKey().toInt() ) );
             elem.setAttribute( "ordered", String.valueOf( entity.isOrderedSection() ) );
         }
